+++ RandomGen/MazeConnector.cpp
@@ -8,7 +8,7 @@ using namespace GameDefs;
 void MazeConnector::ConnectMaze(const std::vector<sf::IntRect>& rooms, std::vector<std::vector<Tile>>& tiles, const GameDefs::GenerateType& genType, unsigned seed, int sleepDuration)
 {
 	m_rooms = rooms;
-	m_rowCount = tiles.size();
+	m_rowCount = (int)tiles.size();
 	if (m_rowCount < 1 || m_rooms.size() < 1)
 		return;
 	{
@@ -20,7 +20,7 @@ void MazeConnector::ConnectMaze(const std::vector<sf::IntRect>& rooms, std::vect
 	m_randomNumGen.seed(seed);
 	m_tiles = &tiles;
 	m_seed = seed;
-	m_columnCount = (*m_tiles)[0].size();
+	m_columnCount = (int)(*m_tiles)[0].size();
 
 	if (genType == Step)
 	{
